remote-desktop: Unify codestyle in API
The code style in the remote-desktop API is currently a mess.
While, the overall code style in mutter and gnome-remote-desktop is to
use spaces as tabulator, the remote-desktop API is mixed with both
spaces as tabulators and actual tabulators.
In addition to that, the code style in the XDG desktop portal APIs uses
spaces as tabulators as well.

To unify the code style, replace all actual tabulators with spaces.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2255>