build: Update sysprof build options
Sysprof's build options have changed recently. This both bumps the sysprof
version and updates the configuration options for the subproject.

If now is not a good time to bump this, that is totally fine, but I wanted
to give you a MR with the necessary changes all in one.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2428>